[.NET] Centralize static methods and object creation
The new Application class combines the previous CanteraInfo and CanteraLogger classes and provides factory methods for creating obejcts such as ThermoPhase. This allows running any required initialization code.
The ApplicationFixture class combined with [Collection("Application")] ensure that tests which need to share state (i.e. use the Application class) do not run in parallel.
