+++ interfaces/dotnet/Cantera.Tests/src/TestFixtures.cs
@@ -1,0 +1,14 @@
+using Xunit;
+
+namespace Cantera.Tests;
+
+public class ApplicationFixture
+{
+    public ApplicationFixture()
+    {        
+        Application.DataDirectories.AddAssemblyDirectory();
+    }
+}
+
+[CollectionDefinition("Application")]
+public class DatabaseCollection : ICollectionFixture<ApplicationFixture> { }
