+++ interfaces/dotnet/Cantera.Tests/src/ThermoPhaseTest.cs
@@ -2,18 +2,17 @@ using Xunit;
 
 namespace Cantera.Tests;
 
+[Collection("Application")]
 public class ThermoPhaseTest
 {
-    static readonly string InputFile = Path.GetFullPath("gri30.yaml");
-
     [Fact]
     public void ThermoPhase_SpeciesRetrieved()
-    {
-        using var thermo = new ThermoPhase(InputFile);
+    {        
+        using var thermo = Application.CreateThermoPhase("gri30.yaml");
 
         Assert.NotEmpty(thermo.Species);
         Assert.NotEmpty(thermo.Species[0].Name);
         Assert.Equal(thermo.Species.Count, thermo.Species.MoleFractions.Length);
         Assert.Equal(thermo.Species.Count, thermo.Species.MassFractions.Length);
     }
-} 
+}
